+++ python/michelangelo/canvas/lib/model/tte/unit_test_model_files/model.safetensors
