+++ python/michelangelo/canvas/lib/model/tte/unit_test_model_files/pytorch_model.bin
