+++ python/michelangelo/canvas/lib/model/tte/unit_test_model_files/script_reduction_model.pt
