+++ python/michelangelo/canvas/lib/shared/utils.py
